tests/requirements.txt
pytest>=7.4.0
pytest-html>=3.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-xdist>=3.3.0
pytest-asyncio>=0.21.0
selenium>=4.15.0
webdriver-manager>=4.0.1
requests>=2.31.0
pandas>=1.5.0
openpyxl>=3.0.9
python-dotenv>=0.19.0
faker>=15.0.0